-- Arcade -- You can no longer take Armor from GameEntities Disabled 1.13 swimming mechanic Player kills/deaths mechanic has been fixed (No longer caps at 1)
Increased ScoreboardManager update frequency ( From 1/2 second to 1/5 second)

Team armor is now unbreakable

Items can now have a Base64 value (Used to apply player skins to heads)

New function to get next block in a particular direction

+++ src/net/blockcade/Moshpit/Variables/Hardpoint.java
@@ -56,7 +56,7 @@ public class Hardpoint {
     public void resetTime() {
         this.time = Main.BASE_POINT_TIME*20;
     }
-    public void tickTime(int i) {
+    public void tickTime(double i) {
         this.time = this.time-i;
     }
 }
